Bug 19206: Include a 128 bit random tag as part of the domain isolator nonce.
When creating a domain isolation nonce, 128 bits of entropy is drawn
from a cryptographic source and saved on a per-domain basis.  The new
circuit behavior is changed to regenerate the nonce, instead of
incrementing a counter.

This allows the "right thing" to happen when the same tor instance is
used across multiple Tor Browser sessions, for example when using a
system wide tor, or a magic anonymity box.