Bug 8725: Block `chrome://` based fingerprinting with nsIContentPolicy.
Most addons do not set `contentaccessible=yes`, however behavior should
be consistent even if such addons are installed.

This does not affect any of the standard addons shipped with Tor Browser, but
will break user installed addons that depend on actually being able to
access `chrome://` URLs in this manner.