Bug 8725: Consistently deny redirects to browser/addon internal URLs.
The browser's behavior is different depending on if a given internal
resource is available or not, regardless of the fact that the actual
body will not load due to the various safeguards and checks.

This normalizes the behavior by denying all redirects destined for URLs
with proscribed browser internal schemes (`resource`, `about`, `chrome`).