Bug 20244.1: Remove "block-disk" checkbox from Privacy & Security Settings
An equivalent checkbox, "Always use private browsing mode"
is already available in about:preferences#privacy. Better to leave
it more hidden from the user in any case, as it is inadvisable
to disable private browsing mode.

We remove the 'extensions.torbutton.block_disk' pref and use the
'browser.privatebrowsing.autostart' pref instead, because they
are equivalent.