Bug 21865: Update our JIT preferences in the slider
It turns out that the JIT *.content prefs are gone for a while now. This
happened in https://bugzilla.mozilla.org/show_bug.cgi?id=939562 and the
patch updates the security slider to take this into account. We got the
tip to include `javascript.options.native_regexp` as well.

`javascript.options.typeinference` is gone with
https://bugzilla.mozilla.org/show_bug.cgi?id=972817 and we therefore
remove it. And asm.js is disabled globally until we find a good solution
for #19417.

The tooltip text got updated accordingly.