Bug 20375: Warn users after entering fullscreen mode.
When the browser window is maximized, we show users a notification
warning that maximizing the screen is fingerprintable. We should
show this warning when the window enters fullscreen mode as well.
This change is especially important for macOS, which uses
window.STATE_FULLSCREEN instead of window.STATE_MAXIMIZED when
the user clicks the "maximize" button on the window title bar.