Bug 26409: Remove spoofed locale implementation
In bug 1039069 Mozilla implemented the option of spoofing the locale in
the Accept-Language header/JS locale if users with localized builds want
that for better fingerprinting protection. That makes our workaround in
Torbutton obsolete and we can remove it. Otherwise users start to get
the respective language prompt twice, which is quite annoying.