Bug 27663: Add New Identity menuitem again
Moving the New Identity button to the hamburger menu without having the
redesign of our security controls in place and without having a proper
guidance for where we moved it to is confusing. Let's revert that
decision until we have all required pieces ready.

We omit the key attribute from the menu item to avoid doing a New
Identity twice if users press Ctrl+Shift+u.