Bug 28515: Use en-US for english Torbutton strings
There is no `en` in Mozilla's universe and we need `en-US` for Torbutton
integration into `tor-browser` (see: #25013). Thus, switch our locale
code for english in Torbutton accordingly.

The server-side of this switch got done in #28261 and the workaround for
that, which was symlinking `en` to `en-US`, is not needed anymore.