Bug 27290: Remove WebGL pref for min capability mode
In Mozilla Bug 1217290, a new WebGL mode was introduced specifically
for `privacy.resistFingerprinting` set to `true`, which is supposed to
give most machines the same WebGL profile.

Previously, `webgl.min_capability_mode` was used for this purpose,
but now that pref is intended strictly for browser testing and it
overrides the `privacy.resistFingerprinting` setting. Furthermore,
`webgl.min_capability_mode` can cause WebGL to fail unnecessarily.

So we should not be enabling `webgl.min_capability_mode` any more in
Tor Browser. This patch is the complement to a tor-browser patch
which both remove that preference setting so it reverts to Firefox's
default value (false).