Bug 28836: Links on about:tor are not clickable.
When the about:tor window was short, links such as the Tor Browser
manual one were not clickable (the fundraising banner made this
problem more likely). The root cause was that the onions/circles
at the bottom of the page overlapped with the links, and the divs
that were used to generate the circles were in front of the links.

This patch changes the behavior on desktop to not allow the onions
to be placed under any page content; instead, they are displayed
at the bottom of the page after all other content. If they won't
fit on the page, a vertical scrollbar is shown.

On mobile, the existing behavior is preserved (onions can appear
behind content). We added `z-index: -1` to ensure that all tappable
content will be usable (although no links are currently displayed
on mobile, we might add some in the future).