Bug 29440: Update about:tor when Tor Browser is updated
After the browser has been updated, notify the user by displaying
a message on the about:tor page. Also, in release and alpha builds,
include on the about:tor page below the Tor Browser version a link
to a simplified changelog page (which continues to be located at
about:tbupdate).