Bug 29825: Intelligently add the new Security Level button to the taskbar
Added logic to torbutton to add the Security Level button after the
url bar, and subsequently move Tor Button between. This new method keeps
any UI customizations the user has made. In the event of error, the UI
configuration reverts back to the default as specified in the
browser.uiCustomization.state pref.